orion/robot-mignon/nas/vision/requirements.txt
# Frameworks généraux
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
httpx==0.27.0

# Traitement d'images et vision par ordinateur
opencv-python==4.8.1.78
numpy==1.26.3
Pillow==10.2.0
scikit-image==0.22.0

# Intelligence artificielle et détection d'objets
torch==2.1.2
torchvision==0.16.2
tensorflow==2.15.0
tensorflow-hub==0.15.0

# Détection d'objets et visages
face-recognition==1.3.0
mtcnn==0.1.1
ultralytics==8.1.0  # YOLOv8

# Intégration avec le serveur MCP
requests==2.31.0
aiohttp==3.9.3

# Traitement des fichiers
python-multipart==0.0.6 